feat(etcd): Enable SSL/TLS for peer-to-peer communication
This commit configures etcd to use SSL/TLS for its peer-to-peer
communication when the `etcd_ssl_enabled` variable is set to `true`.
Client-to-server connections are unaffected and will be handled in a
subsequent change.

To achieve this, a new `etcd_protocol` fact (`http` or `httpss`) has
been introduced to avoid repetitive logic in the configuration
template. The `etcd.conf.j2` template is updated to:

* Use this protocol for all peer-related URLs.
* Conditionally add the `ETCD_PEER_` TLS parameters (cert, key, and CA)
  when SSL is enabled.

References: TPA-1131.

Signed-off-by: Israel Barth Rubio <israel.barth@enterprisedb.com>