feat(patroni): Add `patroni_etcd_user` configuration variable
This commit introduces the `patroni_etcd_user` configuration variable.
This will be used to define the username for a dedicated,
least-privilege user that Patroni will use to connect to etcd when
basic authentication is enabled.

A default value of `etcd_patroni` is provided in the `init` role,
which can be overridden by the user. This change is the first step in
setting up a specific user for Patroni with tightly controlled
permissions in the etcd cluster.

Note: as the inventory used for generating passwords for Ansible
Tower does not support leveraging TPA's config.yml, in that case the
password is always generated for a hard-coded `etcd_patroni` user.
In other words, if the user is on Tower and they configure
`patroni_etcd_user` in their `config.yml` with a value other than
`etcd_patroni`, the deployment will fail.

References: TPA-1136.

Signed-off-by: Israel Barth Rubio <israel.barth@enterprisedb.com>