feat(patroni): Configure etcd basic authentication credentials
This commit completes the etcd basic authentication feature by
configuring the Patroni client with the necessary credentials.

When `etcd_authentication_mode` is `basic`, a new configuration block
is now conditionally merged into the Patroni YAML file. This block
populates the `etcd3` section with the username (from the
`patroni_etcd_user` variable) and the corresponding password.

This enables Patroni to successfully authenticate with and connect to
a secured etcd cluster.

References: TPA-1136.

Signed-off-by: Israel Barth Rubio <israel.barth@enterprisedb.com>