fix(patroni): Enable server certificate validation in API role
Previously, all `uri` calls to the Patroni REST API were made with
`validate_certs: no`. This was a necessary workaround for the
self-signed certificates used by the server, but it left the
communication vulnerable to man-in-the-middle attacks.

Now that the server-side generates proper CA-signed certificates, this
commit updates the `patroni/api` Ansible role logic to perform strict
validation.

The `ca_path` parameter is now enabled in all `uri` tasks, pointing to
the trusted CA certificate. This ensures that all internal API
communication is secure and trusted, resolving the part of the
`CERTIFICATE_VERIFY_FAILED` issue.

References: TPA-449

Signed-off-by: Israel Barth Rubio <israel.barth@enterprisedb.com>