feat(patroni): Configure Patroni to use mTLS credentials for etcd
This commit updates the Patroni configuration to support mTLS
authentication with the etcd cluster.

When `etcd_authentication_mode` is set to `mtls`, the `patroni.yml` file
is now correctly populated with the `etcd3.cert` and `etcd3.key`
parameters. This enables the Patroni agent to present its client
certificate and successfully authenticate with an mTLS-secured etcd
cluster.

References: TPA-1138.

Signed-off-by: Israel Barth Rubio <israel.barth@enterprisedb.com>