refactor(patroni): Use `patroni_compare_config` module for configuration updates
This commit refactors the Patroni configuration task to use the new
`patroni_compare_config` custom Ansible module.

Previously, the logic would always trigger a reload of the Patroni
agent upon configuration changes. However, in some cases a reload is
not enough, and a restart is required.

The new implementation is significantly cleaner and more robust. The
playbook now calls the `patroni_compare_config` module to get the
required `action` and notifies the appropriate handler (reload or
restart).

References: TPA-1138.

Signed-off-by: Israel Barth Rubio <israel.barth@enterprisedb.com>