feat(patroni): Add patroni_authentication_mode variable
This commit introduces a new configuration variable,
`patroni_authentication_mode`, to control the client authentication
method for the Patroni REST API.

The available options are `basic` and `mtls`.

The default is set to `basic` to maintain backward compatibility, as
previous deployments were implicitly configured with username/password
authentication. This ensures that existing clusters do not break upon
upgrade. This new variable formalizes that setting and allows users to
select a different mode.

References: TPA-1135.

Signed-off-by: Israel Barth Rubio <israel.barth@enterprisedb.com>