Add logic for upgrade_components list
Allow for explicit opt-in of which components to upgrade, passed
to the --components flag. This is parsed into a list named
upgrade_components, and the new strict_selects filter ensures
any element is strictly included in this list. If no arguments
are passed to this flag, postgres will automatically be added
and upgrade will only upgrade Postgres by default.

References: TPA-853