Checks for minor version upgrade for PGD-X
Before doing a minor version upgrade of Postgres and PGD, we do cluster
health checks and record the currently running versions of Postgres
and PGD. Then we ensure repositories are updated and check that package
updates won't cause conflicts. After upgrading each host, we check that
postgres and Connection Manager are listening on the expected ports. After
upgrading all hosts, we re-run the health checks and output the original
and currently versions.

A new task file in the "tests" role contains the listening ports check.

References: TPA-1040