Install libcurl-full when required
libcurl-minimal is installed certain RHEL-esque images, and while it
satisfies the package dependencies for edb-pem-agent, it does NOT
provide the complete set of APIs required (namely for SMTP) If it is
installed, we remove it and install the libcurl-full meta-package to
ensure full libcurl functionality exists for the PEM agent. Because
the edb-pem package depends on the edb-pem-agent package, the
libcurl-minimal package must also be removed on the PEM server in
order to avoid dependency resolution conflicts. This is done by
installing the package with the allowerase parameter.

References: TPA-1211