unwanted_packages does not work when choosing the common option
When choosing a package using the common option, this does not remove the unwanted package
from the list provided if that package belonged to the common package list, only when
choosing the specific distribution it belonged it would be removed.
A fix has been introduced where it would skip the installation of the package, or if present,
it will remove it.

References: TPA-1210