Install libpq5 on PEM server v10+ for RHEL
PEM 10.1 added libpq5 as a dependency to fix psycopg errors related
to kerberos authentication on RHEL 9 servers. A postgres-server
package could be installed which already provides libpq5 (as in
providing capability, not always a package with the exact name).
To ensure libpq5 is installed, TPA now explicitly installs it on RHEL
servers when the PEM server version is 10 or greater.

References: TPA-1166