Automatic updates to collections/requirements.yml
The new script check_one_collection.sh takes the name of an ansible
galaxy collection. It looks for the current version of the collection in
collections/requirements.yml and asks the galaxy api for the latest
version; if the latest one is more recent, it writes an updated version
of collections/requirements.yml. It outputs a boolean indicating whether
a change has been made, and a link to the collection's changelog if so.

The new workflow "Ansible collection update" looks at
collections/requirements.yml and calls the update-requirements action
for each collection. If the version has changed, it creates a PR to
update the file, or updates an open PR if there is already one for
the same collection. This workflow is run automatically weekly, on
Monday mornings.

The existing Python requirements file update PR also calls the script
in a loop, adding all the changes to the single PR it generates

References: TPA-850