Support upgrade of components in BDR/PGD clusters
The software component versions can now be updated, either to a
specific version by specifying the component package version in
config.yml or to the latest available when it is not provided. A list
of components for upgrade can be passed to the --components flag as a
comma-separated list. If this flag is not passed, postgres will be
updated. Speciyfing --components=all will update all applicable
software components.

References: TPA-927