Fix race condition on barman_user's pgpass
This change ensures that the .pgpass file on a Barman node used
to back up multiple nodes will not be updated simultaneously by
those backed-up nodes.

Serializing the process avoids a race condition that could cause
only part of the required updates to be saved to the file.

With this change, TPA will force the update to occur one node at a
time, allowing each change to be correctly saved to the .pgpass file
on the Barman node.

Solution provided by support team after testing with the customer.

References: TPA-1207