Cleanup release note leftovers from 23.40.1 release
When performing the 23.40.1 patch release, the release notes that
went out with that release, had leftover release note yaml files.

This commit intends to clean those up so that with the next minor
release, that is 23.41, we will not have duplicate entries for things
released in the previous path release.

References: TPA-1213

Signed-off-by: Martín Marqués <martin.marques@enterprisedb.com>