Support upgrade of components in EFM clusters
The software component versions can now be updated, either to a
specific version by specifying the component package version in
config.yml or to the latest available when it is not provided. A list
of components for upgrade can be passed to the --components flag as a
comma-separated list. If this flag is not passed, postgres will be
updated. Speciyfing --components=all will update all applicable
software components. Consult upgrade documentation, as certain
components only support minor-version upgrades (i.e. Postgres) and
others support both minor and major version upgrades (i.e. EFM, PEM).

References: TPA-926